Fix for bank

+++ public_html/app/models/user.php
@@ -157,6 +157,6 @@ function user_create($email, $passwd, $username, $group = 2)
 function user_helperbalance($balance = 0, $salt)
 {
     $salt = (int)preg_replace('/[^0-9.]+/', '', $salt);
-    return (int)($balance - $salt);
+    return (float)($balance - $salt);
 }
